# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Restaurant Inventory Management API
  description: Comprehensive API for restaurant inventory management, waste tracking, and analytics
  version: 1.0.0
  contact:
    name: Restaurant Inventory Team
    email: support@restaurant-inventory.com
servers:
  - url: https://api.restaurant-inventory.com/v1
    description: Production server
  - url: https://dev-api.restaurant-inventory.com/v1
    description: Development server
paths:
...
  /restaurants/{restaurantId}/inventory/{itemId}:
    get:
      tags: [Inventory]
      summary: Get inventory item
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - $ref: '#/components/parameters/ItemId'
      responses:
        '200':
          description: Item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Inventory]
      summary: Update inventory item
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInventoryItemRequest'
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Inventory]
      summary: Delete inventory item
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - $ref: '#/components/parameters/ItemId'
      responses:
        '204':
          description: Item deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # Real-time Updates
...
components:
  parameters:
    RestaurantId:
      name: restaurantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Restaurant identifier
    ItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Inventory item identifier
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Validation failed"
            details: ["Name is required"]
            timestamp: "2024-01-15T10:30:00Z"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Resource not found"
            timestamp: "2024-01-15T10:30:00Z"
  schemas:
...
    InventoryCategory:
      type: string
      enum: [produce, meat, dairy, dry_goods, beverages, frozen, spices, cleaning]
    InventoryUnit:
      type: string
      enum: [kg, lbs, pieces, liters, gallons, oz, cups]
    InventoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        restaurantId:
          type: string
          format: uuid
        name:
          type: string
          example: "Roma Tomatoes"
        category:
          $ref: '#/components/schemas/InventoryCategory'
        quantity:
          type: number
          minimum: 0
          example: 25.5
        unit:
          $ref: '#/components/schemas/InventoryUnit'
        costPerUnit:
          type: number
          minimum: 0
          example: 2.50
        supplier:
          type: string
          example: "Fresh Produce Co"
        expirationDate:
          type: string
          format: date
          example: "2024-01-20"
        minimumStock:
          type: number
          minimum: 0
          example: 5.0
        isLowStock:
          type: boolean
          readOnly: true
        isExpiringSoon:
          type: boolean
          readOnly: true
        totalValue:
          type: number
          readOnly: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateInventoryItemRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        category:
          $ref: '#/components/schemas/InventoryCategory'
        quantity:
          type: number
          minimum: 0
        unit:
          $ref: '#/components/schemas/InventoryUnit'
        costPerUnit:
          type: number
          minimum: 0
        supplier:
          type: string
          minLength: 1
          maxLength: 255
        expirationDate:
          type: string
          format: date
        minimumStock:
          type: number
          minimum: 0
...
    InventoryItemResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/InventoryItem'
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: string
          example: ["Name is required", "Email must be valid"]
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token obtained from /auth/login endpoint"
security:
  - BearerAuth: []